+++ services/web/src/app/[locale]/(content)/(private)/dashboard/projects/create/page.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import CreateProjectForm from '@/components/project/create-project-form';
-
-export default async function Page() {
-  return <CreateProjectForm />;
-}