+++ services/web/src/app/[locale]/(projects)/layout.tsx
@@ -1,4 +1,5 @@
-import { PropsWithChildren } from 'react';
+import React, { PropsWithChildren } from 'react';
+import Header from '@/components/common/header';
 import { auth } from '@/auth';
 import { redirect } from 'next/navigation';
 
@@ -9,5 +10,11 @@ export default async function Layout({
 
   if (!session) redirect('/connect');
 
-  return <>{children}</>;
+  return (
+    <div className="flex flex-1 flex-col">
+      <Header />
+      {children}
+      {/*<Footer/>*/}
+    </div>
+  );
 }